+++ robotframework-ls/src/robotframework_ls/impl/variable_completions.py
@@ -17,7 +17,6 @@ from robotframework_ls.impl.protocols import (
 from robotframework_ls.impl.text_utilities import normalize_robot_name
 from robotframework_ls.impl.variable_types import (
     VariableFoundFromToken,
-    VariableFoundFromPythonAst,
     VariableFoundFromYaml,
 )
 
@@ -185,44 +184,6 @@ def _collect_resource_imports_variables(
         _collect_global_variables_from_document_context(new_ctx, collector)
 
 
-def _gen_var_from_python_ast(variable_import_doc, collector, node, target):
-    import ast as ast_module
-
-    if isinstance(target, ast_module.Name):
-        varname = target.id
-        if collector.accepts(varname):
-            value = ""
-            try:
-                # Only available for Python 3.8 onwards...
-                end_lineno = getattr(node.value, "end_lineno", None)
-                if end_lineno is None:
-                    end_lineno = node.value.lineno
-
-                # Only available for Python 3.8 onwards...
-                end_col_offset = getattr(node.value, "end_col_offset", None)
-                if end_col_offset is None:
-                    end_col_offset = 99999999
-                value = variable_import_doc.get_range(
-                    node.value.lineno - 1,
-                    node.value.col_offset,
-                    end_lineno - 1,
-                    end_col_offset,
-                )
-            except:
-                log.exception()
-
-            variable_found = VariableFoundFromPythonAst(
-                variable_import_doc.path,
-                target.lineno - 1,
-                target.col_offset,
-                target.lineno - 1,
-                target.col_offset + len(target.id),
-                value,
-                variable_name=varname,
-            )
-            collector.on_variable(variable_found)
-
-
 def _collect_variables_from_variable_import_doc(
     variable_import_doc: IRobotDocument, collector: IVariablesCollector
 ):
@@ -230,19 +191,13 @@ def _collect_variables_from_variable_import_doc(
         if variable_import_doc.path.lower().endswith(".py"):
             python_ast = variable_import_doc.get_python_ast()
             if python_ast is not None:
-                import ast as ast_module
+                from robotframework_ls.impl.variable_completions_from_py import (
+                    collect_variables_from_python_ast,
+                )
 
-                for node in python_ast.body:
-                    if isinstance(node, ast_module.AnnAssign):
-                        _gen_var_from_python_ast(
-                            variable_import_doc, collector, node, node.target
-                        )
-
-                    elif isinstance(node, ast_module.Assign):
-                        for target in node.targets:
-                            _gen_var_from_python_ast(
-                                variable_import_doc, collector, node, target
-                            )
+                collect_variables_from_python_ast(
+                    python_ast, variable_import_doc, collector
+                )
 
         elif variable_import_doc.path.lower().endswith((".yaml", ".yml")):
             dct_contents = variable_import_doc.get_yaml_contents()